refactor: :recycle: move `parse_source()` into own file
