+++ src/seedcase_flower/internals.py
@@ -1,4 +1,4 @@
-"""Helper functions for private use."""
+"""Helper functions for internal use."""
 
 import tomllib
 from dataclasses import dataclass
@@ -6,7 +6,6 @@ from importlib.resources import files
 from itertools import repeat
 from pathlib import Path
 from typing import Any, Callable, Iterable, Optional, TypeVar, Union
-from urllib import parse
 
 from cyclopts.annotations import get_hint_name
 from cyclopts.help import HelpEntry
@@ -17,56 +16,6 @@ from pydantic import BaseModel, Field
 from seedcase_flower.config import Config
 from seedcase_flower.section import Content, Mode, Section
 from seedcase_flower.styles import Style
-
-
-@dataclass(frozen=True)
-class Address:
-    """A source parsed into an actual address."""
-
-    value: str
-    local: bool
-
-
-def _parse_source(source: str) -> Address:
-    split_source = parse.urlsplit(source)
-    if split_source.scheme == "":
-        split_source = split_source._replace(scheme="file")
-    match split_source.scheme:
-        case "file":
-            return _convert_to_path(split_source)
-        case "https":
-            return _convert_to_https(split_source)
-        case "gh" | "github":
-            return _convert_to_github(split_source)
-        case _:
-            raise ValueError(
-                "The source must be either a path to an existing file or "
-                "folder or have one of the following prefixes: `https:`, "
-                "`gh:`, `github:`"
-            )
-
-
-def _convert_to_path(source: parse.SplitResult) -> Address:
-    path = Path(source.path).resolve()
-    if path.is_dir():
-        path /= "datapackage.json"
-    source = source._replace(path=path.as_posix())
-    return Address(value=source.geturl(), local=True)
-
-
-def _convert_to_https(source: parse.SplitResult) -> Address:
-    return Address(value=source.geturl(), local=False)
-
-
-def _convert_to_github(source: parse.SplitResult) -> Address:
-    return Address(
-        value=source._replace(
-            scheme="https",
-            netloc="raw.githubusercontent.com",
-            path=f"/{source.path}/refs/heads/main/datapackage.json",
-        ).geturl(),
-        local=False,
-    )
 
 
 def _format_param_help(entry: HelpEntry) -> str:
