+++ tests/test_cli.py
@@ -11,18 +11,18 @@ from rich.markdown import Markdown
 from seedcase_flower.cli import _CONSOLE_THEME, app
 from seedcase_flower.config import Config
 from seedcase_flower.internals import (
-    Address,
     BuiltSection,
     _get_template_dir,
     _load_sections,
 )
+from seedcase_flower.parse_source import Address
 from seedcase_flower.styles import Style, ViewStyle
 
 
 @pytest.fixture
 def mock_parse_source(mocker):
     """Mock _parse_source to isolate CLI tests from filesystem resolution."""
-    return mocker.patch("seedcase_flower.cli._parse_source")
+    return mocker.patch("seedcase_flower.cli.parse_source")
 
 
 @pytest.fixture
@@ -273,7 +273,7 @@ def test_view_styles_are_one_page():
 
 def test_view_with_mocked_internals(mocker):
     """view should parse source, build sections, and render via Console."""
-    mock_parse_source = mocker.patch("seedcase_flower.cli._parse_source")
+    mock_parse_source = mocker.patch("seedcase_flower.cli.parse_source")
     mock_read_properties = mocker.patch("seedcase_flower.cli.read_properties")
     mock_build_sections = mocker.patch("seedcase_flower.cli._build_sections")
     mock_console_cls = mocker.patch("seedcase_flower.cli.Console")
